+++ poem-openapi/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.1.1] 2021-12-12
+
+- Add `BinaryStream::from_bytes_stream` and `BinaryStream::to_bytes_stream` methods.
+
 # [1.0.51] 2021-12-12
 
 - Add some methods to specify more API metadata.