+++ poem/CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [1.0.39] 2021-12-13
+# [1.1.1] 2021-12-13
+
+- Add `Body::from_bytes_stream` and `Body::to_bytes_stream` methods.
+
+# [1.1.0] 2021-12-13
 
 - Remove `nom` from dependencies.